requirements.txt
amqp==5.0.9
aniso8601==9.0.1
attrs==21.4.0
billiard==3.6.4.0
Brotli==1.0.9
celery==5.2.3
cffi==1.15.0
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cssselect2==0.4.1
Deprecated==1.2.13
Flask==2.0.3
flask-marshmallow==0.14.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
Flask-WeasyPrint==0.6
fonttools==4.29.1
greenlet==1.1.2
html5lib==1.1
importlib-resources==5.4.0
itsdangerous==2.1.0
Jinja2==3.0.3
jsonschema==4.4.0
kombu==5.2.3
MarkupSafe==2.1.0
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
packaging==21.3
Pillow==9.0.1
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
pycparser==2.21
pydyf==0.1.2
pyparsing==3.0.7
pyphen==0.12.0
pyrsistent==0.18.1
python-dotenv==0.19.2
python-http-client==3.3.6
pytz==2021.3
redis==4.1.4
sendgrid==6.9.6
six==1.16.0
SQLAlchemy==1.4.31
starkbank-ecdsa==2.0.3
tinycss2==1.1.1
vine==5.0.0
wcwidth==0.2.5
weasyprint==54.1
webencodings==0.5.1
Werkzeug==2.0.3
wrapt==1.13.3
zipp==3.7.0
zopfli==0.1.9
